requirements.txt
google-generativeai>=0.8.0
jsonschema>=4.21.0
sqlparse>=0.5.0
python-dotenv>=1.0.0
pydantic>=2.0.0
chainlit>=1.0.0
openai>=1.0.0
psycopg[binary]>=3.2.0
flask>=2.3.0
